feat(refs: DPLAN-17918): reduce code duplication
