+++ tests/frontend/unit/MultiselectCustomField.spec.js
@@ -8,6 +8,7 @@
  */
 
 import { afterEach, beforeEach, describe, expect, it } from '@jest/globals'
+import { sharedEditableModeTests, sharedEmptyValueTests, sharedMultiselectBasedEditableModeTests, sharedReadonlyModeTests } from './helpers/customFieldSharedTests'
 import { enableAutoUnmount } from '@vue/test-utils'
 import MultiselectCustomField from '@DpJs/components/customFields/MultiselectCustomField'
 import shallowMountWithGlobalMocks from '@DpJs/VueConfigLocal'
@@ -79,41 +80,11 @@ describe('MultiselectCustomField', () => {
       expect(optionDivs[1].text()).toBe('B')
     })
 
-    it('should show a required marker when isRequired is true', () => {
-      const requiredField = {
-        ...defaultField,
-        attributes: { ...defaultField.attributes, isRequired: true },
-      }
-      wrapper = shallowMountWithGlobalMocks(MultiselectCustomField, {
-        props: { field: requiredField },
-      })
-
-      expect(wrapper.find('dt').text()).toContain('*')
-    })
-
-    it('should not show a required marker when isRequired is false', () => {
-      expect(wrapper.find('dt').text()).not.toContain('*')
-    })
-
-    it('should render DpContextualHelp when description is set', () => {
-      const fieldWithDesc = {
-        ...defaultField,
-        attributes: { ...defaultField.attributes, description: 'Help text' },
-      }
-      wrapper = shallowMountWithGlobalMocks(MultiselectCustomField, {
-        props: { field: fieldWithDesc },
-      })
-
-      expect(wrapper.findComponent({ name: 'DpContextualHelp' }).exists()).toBe(true)
-    })
-
-    it('should not render DpContextualHelp when description is empty', () => {
-      expect(wrapper.findComponent({ name: 'DpContextualHelp' }).exists()).toBe(false)
-    })
-
     it('should not render DpMultiselect', () => {
       expect(wrapper.findComponent({ name: 'DpMultiselect' }).exists()).toBe(false)
     })
+
+    sharedReadonlyModeTests(MultiselectCustomField, defaultField)
   })
 
   describe('readonly mode – empty value', () => {
@@ -129,17 +100,7 @@ describe('MultiselectCustomField', () => {
       expect(wrapper.find('dd').text()).toContain('-')
     })
 
-    it('should show a dash when field.value is null', () => {
-      const nullValueField = {
-        ...defaultField,
-        value: null,
-      }
-      wrapper = shallowMountWithGlobalMocks(MultiselectCustomField, {
-        props: { field: nullValueField },
-      })
-
-      expect(wrapper.find('dd').text()).toContain('-')
-    })
+    sharedEmptyValueTests(MultiselectCustomField, defaultField)
   })
 
   describe('editable mode', () => {
@@ -157,34 +118,8 @@ describe('MultiselectCustomField', () => {
       expect(multiselect.props('value')).toEqual([{ id: 'a', label: 'A' }])
     })
 
-    it('should render DpLabel when showLabel is true', () => {
-      expect(wrapper.findComponent({ name: 'DpLabel' }).exists()).toBe(true)
-    })
-
-    it('should not render DpLabel when showLabel is false', () => {
-      wrapper = shallowMountWithGlobalMocks(MultiselectCustomField, {
-        props: { field: defaultField, mode: 'editable', showLabel: false },
-      })
-
-      expect(wrapper.findComponent({ name: 'DpLabel' }).exists()).toBe(false)
-    })
-
-    it('should pass isRequired to DpMultiselect and DpLabel', () => {
-      const requiredField = {
-        ...defaultField,
-        attributes: { ...defaultField.attributes, isRequired: true },
-      }
-      wrapper = shallowMountWithGlobalMocks(MultiselectCustomField, {
-        props: { field: requiredField, mode: 'editable' },
-      })
-
-      expect(wrapper.findComponent({ name: 'DpMultiselect' }).props('required')).toBe(true)
-      expect(wrapper.findComponent({ name: 'DpLabel' }).props('required')).toBe(true)
-    })
-
-    it('should not render a definition list', () => {
-      expect(wrapper.find('dl').exists()).toBe(false)
-    })
+    sharedEditableModeTests(MultiselectCustomField, defaultField)
+    sharedMultiselectBasedEditableModeTests(MultiselectCustomField, defaultField)
   })
 
   describe('update:value emit', () => {
